+++ packages/backend/src/server/api/endpoints/users/notes.ts
@@ -41,8 +41,8 @@ export const paramDef = {
 	type: 'object',
 	properties: {
 		userId: { type: 'string', format: 'misskey:id' },
-		includeReplies: { type: 'boolean', default: true },
-		includeRenotes: { type: 'boolean', default: true },
+		withReplies: { type: 'boolean', default: false },
+		withRenotes: { type: 'boolean', default: true },
 		limit: { type: 'integer', minimum: 1, maximum: 100, default: 10 },
 		sinceId: { type: 'string', format: 'misskey:id' },
 		untilId: { type: 'string', format: 'misskey:id' },
@@ -115,11 +115,11 @@ export default class extends Endpoint<typeof meta, typeof paramDef> { // eslint-
 				}
 			}
 
-			if (!ps.includeReplies) {
+			if (!ps.withReplies) {
 				query.andWhere('note.replyId IS NULL');
 			}
 
-			if (ps.includeRenotes === false) {
+			if (ps.withRenotes === false) {
 				query.andWhere(new Brackets(qb => {
 					qb.orWhere('note.renoteId IS NULL');
 					qb.orWhere(new Brackets(qb => {